Add people and reverse lookup provider "Auskunft" (AT).
Change-Id: Ib57a4aca35bcd80f320c5a2a44faac4c9c9d3026

Lookup: improve "Auskunft" (AT) API.

This is a follow-up on Ib57a4aca35bcd80f320c5a2a44faac4c9c9d3026
which improves the parsing process by working the search response of
section by section to not miss any incomplete data sets.
The old process was kind of error prone because it searched for all
names, numbers and addresses and tried to match those which of
course failed if there was e.g. an address missing.

Change-Id: Ib685477916e06e63e9b63f2fff7f2bf47a9d9ba3
Signed-off-by: Louis Popi <theh2o64@gmail.com>