Remove usage legacy HTTP Client
Improved the "LookupUtils" to not use the deprecated HTTP classes
but "HttpURLConnection" instead. Contrary to the Apache HTTP client
the HttURLConnection is still existing in Android 6 and seems to be
more efficient besides being compatible with SSL SNI.

Furthermore set the missing "formattedNumber" attribute when the
"ContactBuilder" is creating the "ContactInfo" object after lookups.

Change-Id: I6b1e58ab3a1f18804cdaeb8031805230de45de5a
Signed-off-by: Louis Popi <theh2o64@gmail.com>